pyproject.toml
[project]
name = "sono"
version = "0.1.0"
description = "Sono App"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Dennis Paz", email = "dppazlopez@gmail.com" },
  { name = "Alisson Pereira", email = "alissonpef@gmail.com" },
]
requires-python = "==3.14.*"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "argon2-cffi==25.1.0",
  "celery==5.6.3",
  "crispy-tailwind==1.0.3",
  "django==6.0.4",
  "django-allauth[mfa]==65.16.1",
  "django-anymail[amazon-ses]==15.0",
  "django-braces==1.17.0",
  "django-celery-beat==2.9.0",
  "django-cotton==2.6.2",
  "django-crispy-forms==2.6",
  "django-environ==0.13.0",
  "django-extra-views==0.16.0",
  "django-filter==25.2.0",
  "django-formtools==2.5.1",
  "django-htmx==1.27.0",
  "django-jsonform==2.23.2",
  "django-model-utils==5.0.0",
  "django-phonenumber-field[phonenumberslite]==8.4.0",
  "django-redis==6.0.0",
  "django-storages[s3]==1.14.6",
  "django-vite==3.1.0",
  "django-widget-tweaks==1.5.1",
  "gunicorn==25.3.0",
  "hiredis==3.3.1",
  "pillow==12.2.0",
  "psycopg[c]==3.3.3",
  "python-slugify==8.0.4",
  "redis==7.4.0",
  "whitenoise==6.12.0",
]

[dependency-groups]
dev = [
  "django-browser-reload==1.21.0",
  "django-debug-toolbar==6.3.0",
  "django-extensions==4.1",
  "django-watchfiles==1.4.0",
  "djlint==1.36.4",
  "factory-boy==3.3.3",
  "ipdb==0.13.13",
  "pre-commit==4.6.0",
  "psycopg[binary]==3.3.3",
  "pytest==9.0.3",
  "pytest-django==4.12.0",
  "pytest-sugar==1.1.1",
  "ruff==0.15.12",
  "tailwhip==0.13.0",
  "watchfiles==1.1.1",
  "werkzeug[watchdog]==3.1.8",
]

[tool.ruff]
extend-exclude = [
  "*/migrations/*.py",
  "staticfiles/*",
]
lint.select = [
  "A",
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "DJ",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FBT",
  "FLY",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "N",
  "PD",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  "COM812", # Conflicts with Ruff formatter
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "SIM102", # sometimes it's better to nest of types for comparison
]
lint.isort.force-single-line = true

[tool.pyproject-fmt]
keep_full_version = true

[tool.pytest]
ini_options.minversion = "6.0"
ini_options.addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
ini_options.python_files = [ "tests.py", "test_*.py" ]

[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"
custom_html = "c-\\w+,c-\\w+-\\w+,c-\\w+-\\w+-\\w+,c-\\w+-\\w+-\\w+-\\w+,c-\\w+-\\w+-\\w+-\\w+-\\w+,circle,rect,path,svg,use,line"
custom_blocks = "cotton,partialdef,slot"
css.indent_size = 2
js.indent_size = 2

[tool.tailwhip]
default_globs = [ "templates/**/*.html", "static/css/**/*.css" ]
custom_colors = [
  "primary",
  "secondary",
  "accent",
  "neutral",
  "base",
  "info",
  "success",
  "warning",
  "error",
]
